login/src/main/java/com/bci/login/service/UserService.java
package com.bci.login.service;

import com.bci.login.exception.UserAlreadyExistsException;
import com.bci.login.model.User;
import com.bci.login.repository.UserRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.Optional;
import java.util.UUID;

@AllArgsConstructor
@Service
public class UserService {

    private final UserRepository userRepository;

    public User saveUser(User user) {
        if (userRepository.findByEmail(user.getEmail()).isPresent()) {
            throw new UserAlreadyExistsException("Usuario con mail " + user.getEmail() + " ya existe.");
        }
        return userRepository.save(user);
    }

    public Optional<User> getUserById(UUID id) {
        return userRepository.findById(id);
    }
}
